Rename ASTClassesList.h to uast-classes-list.h
Would like to use dashed style for things that aren't
a particular class.

---
